Update to catch exceptions
A few things were wrong with the first attempt. First, there was no Try Except block.  The float would immediately cause an error if the input was non-numeric.  Now, if a non-numeric input is entered, the result will be 'Bad score'.  Second, the line for an F score was redundant in having <1.0 and <0.6.  The <1.0 is no longer necessary since a number less than 0.6 is already less than 1.0.  Third, it's necessary to ensure the entry is greater than 0 to remain in range, because otherwise a negative input would count as an F.